driver/led: add brightness config to the interface of blink
For the original designation of blink interface, the parameter of blinking
is not set to the value of brightness. It just check if blinking is 0. If so,
it will turn off led blink. Or it will set brightness as max value for blink.
It is not suitable for debug. And the brightness of blink could not be
changed. So we changed its logic. To let it change blink brightness.
The meaning of parameter blinking is as following:
0 - turn off blink
1- turn on blink as max brightness(used for driver original designation)
other value - set brightness as desired value for blinking.

Change-Id: Ibc27b6161372c6a0a88174eb2a7b697f00ab23e4
Signed-off-by: weiweij <weiweij@motorola.com>
Reviewed-on: https://gerrit.mot.com/1312998
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key